Crawler working with loop variables
