+++ brds/core/crawler/root_crawler.py
@@ -1,90 +1,11 @@
-from copy import deepcopy
-from itertools import product
-from os.path import join
-from typing import Any, Dict, Iterable, List, Tuple
-from urllib.parse import urlparse
-
-from brds.core.crawler.browser_emulator import BrowserEmulator
-from brds.core.crawler.config import ConfigStore, remove_default_params
+from brds.core.crawler.config import remove_default_params
+from brds.core.crawler.crawler import Crawler
 from brds.core.crawler.variables import VariableHolder
-from brds.core.fs.writer import FileWriter
-from brds.db.init_db import Database
+from brds.core.crawler.templated_url import TemplatedUrl, get_path_from_url
+from brds.core.logger import get_logger as _get_logger
 
 
-class Crawler:
-    def __init__(
-        self: "Crawler",
-        configs: ConfigStore,
-        database: Database,
-        browser_emulator: BrowserEmulator,
-        file_writer: FileWriter,
-        name: str,
-        variables: List[str],
-        inputs: List[str],
-        urls: List[Dict[str, Any]],
-        loop_variables: List[str],
-        _filepath: str,
-    ) -> None:
-        self.configs = configs
-        self.database = database
-        self.browser_emulator = browser_emulator
-        self.file_writer = file_writer
-
-        self.name = name
-        self.variables = variables
-        self.inputs = inputs
-        self.urls = urls
-        self.loop_variables = loop_variables
-        self._filepath = _filepath
-
-    def execute(self: "Crawler") -> None:
-        for input_variables in self.iterate_inputs():
-            vars = self.merge_variables(input_variables)
-            orig_vars = deepcopy(vars)
-            print("ORIG:", orig_vars.variables)
-            for loop_vars in self.iterate_loop_variables(orig_vars):
-                for key, value in zip(self.loop_variables, loop_vars):
-                    if key not in ["name", "_filepath"]:
-                        vars[key] = value
-                print("CALLED:", vars.variables)
-                self._process(vars)
-
-    def merge_variables(self: "Crawler", input_variables: Tuple[Dict[str, Any]]) -> VariableHolder:
-        variables = VariableHolder()
-        variables.extend(
-            {
-                "name": self.name,
-                "_filepath": self._filepath,
-            }
-        )
-        for input in input_variables:
-            variables.extend(remove_variable_parameters(input))
-        for variable in self.variables:
-            variables.extend(remove_variable_parameters(self.configs[variable]))
-        return variables
-
-    def iterate_inputs(self: "Crawler") -> Iterable[Tuple[Dict[str, Any]]]:
-        return product(*[self.configs.get_by_type(input) for input in self.inputs])
-
-    def iterate_loop_variables(self: "Crawler", variables: VariableHolder) -> Iterable[Tuple[str]]:
-        return product(*[variables[loop_variable] for loop_variable in self.loop_variables])
-
-    def _process(self: "Crawler", variables: VariableHolder) -> None:
-        self.process(variables)
-
-    def process(self: "Crawler", variables: VariableHolder) -> None:
-        raise NotImplementedError("You need to override this function")
-
-    def url(self: "Crawler", variables: VariableHolder) -> str:
-        return variables["url"] + self.urls[0]["url"].format(**variables.variables)
-
-
-def remove_variable_parameters(params: Dict[str, Any]) -> Dict[str, Any]:
-    copy = deepcopy(params)
-    for key in ["name", "_filepath"]:
-        if key in copy:
-            del copy[key]
-    return copy
+LOGGER = _get_logger()
 
 
 class RootCrawler(Crawler):
@@ -97,26 +18,22 @@ class RootCrawler(Crawler):
     def process(self: "RootCrawler", variables: VariableHolder) -> None:
         for templated_url in self.templated_urls:
             url = templated_url.resolve(variables)
-            url_id = self.database.get_url_id(url)
-            assert url_id is not None
+            url_id = self.database.register_web_page(url)
             self.database.set_vriables(url_id, variables.variables)
-            if self.should_load(url, templated_url.cache):
-                self.download(url)
+            if self.should_load(url_id, templated_url.cache):
+                self.download(url, url_id)
             else:
-                print(f"Will not download '{url}', as I've already downloaded it")
+                LOGGER.info(f"Will not download '{url}', as I've already downloaded it")
 
-    def should_load(self: "RootCrawler", url: str, cache: bool) -> bool:
+    def should_load(self: "RootCrawler", url_id: int, cache: bool) -> bool:
         if not cache:
             return True
-        url_id = self.database.register_web_page(url)
         last_crawl = self.database.latest_download(url_id)
         return not last_crawl
 
-    def download(self: "RootCrawler", url: str) -> None:
-        url_id = self.database.get_url_id(url)
-        assert url_id is not None
+    def download(self: "RootCrawler", url: str, url_id: int) -> None:
         file_path = get_path_from_url(url)
-        print(f"Downloading '{url}' to '{file_path}'")
+        LOGGER.info(f"Downloading '{url}' to '{file_path}'")
 
         response = self.browser_emulator.get(url)
         full_path = self.file_writer.write(file_path, response)
@@ -128,29 +45,3 @@ class RootCrawler(Crawler):
             str(full_path),
             response.status_code,
         )
-
-
-class TemplatedUrl:
-    def __init__(self: "TemplatedUrl", database: Database, name: str, url: str, cache: bool) -> None:
-        self.name = name
-        self.url = url
-        self.cache = cache
-
-    def resolve(self: "TemplatedUrl", variables: VariableHolder) -> str:
-        return variables["base_url"] + self.url.format(**variables.variables)
-
-
-def sanitize_component(component: str) -> str:
-    return "".join(c if c.isalnum() or c in "-_." else "_" for c in component)
-
-
-def get_path_from_url(url: str) -> str:
-    parsed = urlparse(url)
-
-    domain_path = join(*sanitize_component(parsed.netloc).split("."))
-
-    path = parsed.path if parsed.path else "/"
-    path_components = [sanitize_component(component) for component in path.strip("/").split("/")]
-
-    base_path = join(domain_path, *path_components)
-    return base_path
